requirements.txt
django
psycopg2-binary
python-dotenv
requests
pillow